Linked UI elements on home page to Live data from room. Also managed to link a percentage value to the circle progress bar.

+++ app/src/main/java/com/example/chronoboss/database/Day.kt
@@ -9,7 +9,7 @@ data class Day (
     //Starting with dayId to get it to work. Changing it to a date representing a single day later.
     @PrimaryKey(autoGenerate = true)
     val dayId: Int,
-    val timeWasted: Int,
-    val timeLimit: Int,
+    val timeWasted: Long,
+    val timeLimit: Long,
     val application: String,
         )
